Move hero to index
