+++ components/about/HeroImage.module.css
@@ -1,0 +1,22 @@
+.heroImage {
+    display: block;
+    margin: 2em 0;
+    object-fit: cover;
+    background-position: center;
+    background-size: cover;
+    
+    box-shadow: 0 3px 10px #eee;
+    background-color: #eee;
+}
+
+@media (min-width: 700px) {
+    .heroImage {
+        height: 21em;
+    }
+}
+
+@media (max-width: 700px) {
+    .heroImage {
+        height: 15em;
+    }
+}
