+++ components/about/HeroImage.tsx
@@ -1,0 +1,9 @@
+import css from './HeroImage.module.css'
+
+export function HeroImage(props: {src: string}) {
+    return <div
+        className={css.heroImage}
+        style={{
+            backgroundImage: `url(${props.src})`,
+        }}></div>
+}
