+++ components/about/Section.module.css
@@ -1,0 +1,4 @@
+.section {
+    margin-top: 4em;
+    margin-bottom: 6em;
+}
