+++ components/about/Section.tsx
@@ -1,0 +1,5 @@
+import css from './Section.module.css'
+
+export function Section(props: React.PropsWithChildren<{}>) {
+    return <div className={css.section}>{props.children}</div>
+}
