+++ pages/_app.module.css
@@ -1,0 +1,4 @@
+.container {
+    max-width: 800px;
+    margin: 0 auto;
+}
