+++ pages/about/index.module.css
@@ -1,6 +1,4 @@
 .container {
-    max-width: 800px;
-    margin: 0 auto;
     padding: 0 1em;
 }
 
@@ -29,32 +27,4 @@
 .container td {
     border-color: #eee;
     padding: 0.5em 1em;
-}
-
-.section {
-    margin-top: 4em;
-    margin-bottom: 6em;
-}
-
-.heroImage {
-    display: block;
-    margin: 2em 0;
-    object-fit: cover;
-    background-position: center;
-    background-size: cover;
-    
-    box-shadow: 0 3px 10px #eee;
-    background-color: #eee;
-}
-
-@media (min-width: 700px) {
-    .heroImage {
-        height: 21em;
-    }
-}
-
-@media (max-width: 700px) {
-    .heroImage {
-        height: 15em;
-    }
 }
