+++ public/static/locales/en-US/common.json
@@ -1,6 +1,5 @@
 {
     "title": "Kiganix, G.K. (合同会社キガニックス)",
-    "heading": "Kiganix, G.K.",
     "label-for-company-legal-name": "Legal name of company",
     "company-legal-name": "Kiganix, G.K. (合同会社キガニックス)",
     "label-for-company-number": "Corporate number",
