+++ public/static/locales/ja-JP/common.json
@@ -1,6 +1,5 @@
 {
     "title": "合同会社キガニックス (Kiganix, G.K.)",
-    "heading": "Kiganix, G.K.",
     "label-for-company-legal-name": "商号",
     "company-legal-name": "合同会社キガニックス",
     "label-for-company-number": "法人番号",
